Run tests on recurring schedule (monthly)
OSS libraries and tools that have infrequent activity can be subject to
atrophy because the test suite isn't run unless a PR is opened.

To protect against this, we can schedule the test suite to run monthly.
(The date and time is unimportant. Chosen here as a random minute and
day so as to avoid the peak times of "first hour of day" and "first day
of month".)

This will help ensure that any deprecations (such as action runners
themselves, or dependencies, etc) can be caught earlier instead of
"whenever the next pull-request is opened".